requirements2.txt
asgiref @ file:///croot/asgiref_1724072470684/work
Bottleneck @ file:///croot/bottleneck_1731058641041/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
chardet==3.0.4
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama @ file:///work/perseverance-python-buildout/croot/colorama_1698847155966/work
contourpy @ file:///croot/contourpy_1732540045555/work
cryptography==44.0.0
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Django @ file:///croot/django_1731615954020/work
filelock @ file:///work/perseverance-python-buildout/croot/filelock_1701733993137/work
fonttools @ file:///croot/fonttools_1737039080035/work
fsspec @ file:///croot/fsspec_1736274672181/work
gmpy2 @ file:///croot/gmpy2_1738085463648/work
googletrans==4.0.0rc1
gTTS==2.5.4
h11 @ file:///tmp/build/80754af9/h11_1594058940538/work
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.28.1
hyperframe==5.2.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
Jinja2 @ file:///croot/jinja2_1737760107953/work
joblib @ file:///croot/joblib_1718217211762/work
kiwisolver @ file:///croot/kiwisolver_1737039087198/work
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.0
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
networkx @ file:///work/perseverance-python-buildout/croot/networkx_1698865062738/work
nltk==3.9.1
numexpr @ file:///croot/numexpr_1730215937391/work
numpy @ file:///croot/numpy_and_numpy_base_1725470312869/work/dist/numpy-2.0.1-cp312-cp312-linux_x86_64.whl#sha256=e2374991344fa2241a2153ef3d550d3a1cd2d50cb1f1d51eddc82870abc88021
packaging @ file:///croot/packaging_1734472117206/work
pandas @ file:///croot/pandas_1732735089971/work/dist/pandas-2.2.3-cp312-cp312-linux_x86_64.whl#sha256=57b66702d418720ec8483f7c4ec7c08d41815316ad7ce09d5b7bbc34eefcfdfd
pdfminer==20191125
pdfminer.six==20240706
pillow @ file:///croot/pillow_1721059439630/work
ply @ file:///work/perseverance-python-buildout/croot/ply_1698846188635/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.21.0
pyparsing @ file:///croot/pyparsing_1731445506121/work
PyPDF2==3.0.1
PyQt5==5.15.10
PyQt5_sip @ file:///croot/pyqt-split_1736540531116/work/pyqt_sip
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
pytesseract==0.3.13
python-dateutil @ file:///croot/python-dateutil_1716484871790/work
pytz @ file:///croot/pytz_1713974312559/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
regex @ file:///croot/regex_1736540786412/work
requests @ file:///croot/requests_1730999120400/work
rfc3986==1.5.0
safetensors==0.5.2
scipy @ file:///croot/scipy_1737122901252/work/dist/scipy-1.15.1-cp312-cp312-linux_x86_64.whl#sha256=7b1b0001c943e119509d010e5cbb4950dd6e325a83ca2510a25f139dac6d66a8
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
sentencepiece @ file:///croot/sentencepiece-split_1722874460693/work/python
setuptools==75.8.0
sip @ file:///croot/sip_1736541134699/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
sqlparse @ file:///croot/sqlparse_1734704446487/work
sympy @ file:///croot/sympy_1738108488918/work
textblob==0.19.0
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado @ file:///croot/tornado_1733960490606/work
tqdm @ file:///croot/tqdm_1738943501192/work
transformers==4.48.3
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata @ file:///croot/python-tzdata_1690578112552/work
unicodedata2 @ file:///croot/unicodedata2_1736541023050/work
urllib3 @ file:///croot/urllib3_1727769808118/work
wheel==0.45.1
youtube_transcript_api
streamlit
google-generativeai
python-dotenv
pathlib